replace floats with double tracker

+++ include/Tracker.h
@@ -25,20 +25,20 @@ public:
     bool setTrackingMode(TrackingMode mode);
 
     // Public methods for setting targets
-    void setTLE(const String &newTLE);       // Set TLE for satellite tracking
-    void setEquatorial(float ra, float dec); // Set equatorial coords
-    void setGalactic(float l, float b);      // Set galactic coords
+    void setTLE(const String &newTLE);         // Set TLE for satellite tracking
+    void setEquatorial(double ra, double dec); // Set equatorial coords
+    void setGalactic(double l, double b);      // Set galactic coords
 
     void updateTargetCoordinates();
 
 private:
     SemaphoreHandle_t positionMutex;
     TimerHandle_t trackingTimer;
-    float targetAz, targetEl;
+    double targetAz, targetEl;
 
     TrackingMode currentMode;
     bool target_change_flag;
-    float ra, dec, l, b;
+    double ra, dec, l, b;
     String tle;
 
     // Constructors (singleton)
@@ -47,14 +47,14 @@ private:
     Tracker &operator=(const Tracker &) = delete; // Delete assignment operator
 
     // Method to update target coordinates based on current mode
-    void updateFromTLE(float &az, float &el);
-    void updateFromGalactic(float &az, float &el);
+    void updateFromTLE(double &az, double &el);
+    void updateFromGalactic(double &az, double &el);
 
     // Method to check if position is valid
-    bool isValidPosition(float az, float el);
+    bool isValidPosition(double az, double el);
 
     // Method to check if movement is needed
-    bool needsMovement(float az, float el);
+    bool needsMovement(double az, double el);
 
     // Timer to periodically update coordinates
     void setupTrackingTimer();
